debug-1.c: Use int, not long, in union with vector.
2004-07-23  Janis Johnson  <janis187@us.ibm.com>

	* gcc.dg/vmx/debug-1.c: Use int, not long, in union with vector.
	* gcc.dg/vmx/varargs-1.c: Use int, not long, in union with vector.

From-SVN: r85086

+++ gcc/testsuite/gcc.dg/vmx/varargs-1.c
@@ -13,10 +13,10 @@ void printx(T a)
 {
   union {
     T v;
-    unsigned long a[4];
+    unsigned int a[4];
   } u;
   u.v = a;
-  printf("%ld, %ld, %ld, %ld\n", u.a[0], u.a[1], u.a[2], u.a[3]);
+  printf("%d, %d, %d, %d\n", u.a[0], u.a[1], u.a[2], u.a[3]);
 }
 
 void f1(int a, ...)
